+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Docker Compose file to set up local dev environment
 - Updated README steps for setting up local dev environment 
+- Script to reindex ES works into new ES cluster
 ### Fixed
 - Updated unit tests for developmentSetup process
 - Removed front end configurations from tugboat